fixing up down

+++ js/classes/hole.js
@@ -22,18 +22,22 @@ class Hole{
   }
 
   peep(){
-    const wait = Math.round(Math.random() * (3000 - 1000) + 1000); 
+    // const wait = Math.round(Math.random() * (3000 - 1000) + 1000); 
     const hold = this.randomTime();
-    setTimeout(()=>{
+    if (!this.up){
+      // setTimeout(()=>{
       this.lift();
       console.log(`${this.name} is up. hit it now (up: ${this.up})`)
-    }, wait)
-    if (this.up){
+      // }, wait)
       setTimeout(() => {
-        this.drop();
-        console.log(`${this.name} is down. you missed it (up: ${this.up})`)
-        this.level.game.updateCurrHP(-1) //TODO: health to be lost based on baddie in the hole
+        if (this.up){
+          this.drop();
+          console.log(`${this.name} is down. you missed it (up: ${this.up})`)
+          this.level.game.updateCurrHP(-1) //TODO: health to be lost based on baddie in the hole
+        }
       }, hold)
+      console.log("pustide setTimeout")
+      
     }
   }
